requirements.txt
django
djangorestframework
pika
torch
transformers
redis
django-redis